🤫 Don't require `t.Helper` in functions passed to `synctest.Test`
Go 1.25 made the `testing/synctest` package generally available. The
package introduces the `synctest.Test` function, to be used like this:

```go
func TestMyFeature(t *testing.T) {
	synctest.Test(t, func(t *testing.T) {
		// Do stuff that uses goroutines
	})
}
```

Sadly, `thelper` triggers if the anonymous function passed to
`synctest.Test` doesn't call `t.Helper()`. The issue is that this
anonymous function isn't a helper, it's more like a sub-test.

This PR extends `thelper` to grant an exception to functions passed to
`synctest.Test`, just like functions passed to `t.Run` or `b.Run`.

This requires updating to Go 1.25, so this PR is stacked on top of #50.